Small tweaks to links in 04.Lessons_Declarations
Under the bullet point "Read the documentation on [method
declarations](...)", the first link points to something under the header
"Old documentation", so I suspect that link is not quite right. I
updated it to another link that I think it is supposed to point to
(partly because it is the top header for the following link the same
bullet point mentions). As another minor nit while I am there, I took
off the highlighting for the second link.

These changes might not be right, so feel free to reject this PR if they
are not right.